turns out i don't need icon manager
